requirement.txt
pydantic[dotenv]
httpx
devtools
pytest-asyncio
pytest-cov